feat: Added event for attack and draw in Fighter contract
Added `event Attack(address pair, uint256 value)` and `event Draws(address pair, address pair2)` to the `Fighter` contract. These events are emitted when an attack is performed and when a draw occurs during a fight. The `Attack` event includes the address of the attacking pair and the value of the attack. The `Draws` event includes the addresses of both pairs involved in the draw.

Also made some changes to the contract logic by adding private functions `_checkWinner` and `_declareResult`, and updating the `request` and `fight` functions.